+++ src/features/advanced/cxpc-default-1y.ts
@@ -1,5 +1,9 @@
+import { cxpcStore } from '@src/infrastructure/prun-api/data/cxpc';
+import { cxobStore } from '@src/infrastructure/prun-api/data/cxob';
+
 function onTileReady(tile: PrunTile) {
-  subscribe($$(tile.anchor, C.ChartContainer.settings), settings => {
+  subscribe($$(tile.anchor, C.ChartContainer.settings), async settings => {
+    await waitInitialCxpcResponse(tile.parameter!);
     const radioGroups = _$$(settings, C.RadioGroup.container);
     if (radioGroups.length === 0) {
       return;
@@ -12,6 +16,20 @@ function onTileReady(tile: PrunTile) {
     }
     const yearButton = rangeButtons[rangeButtons.length - 1];
     yearButton?.click();
+  });
+}
+
+// The 1y request will not be sent if the initial response has not arrived yet.
+async function waitInitialCxpcResponse(ticker: string) {
+  const broker = computed(() => cxobStore.getByTicker(ticker));
+  await new Promise<void>(resolve => {
+    const check = (data: PrunApi.CXBrokerPrices) => {
+      if (data.brokerId === broker.value?.id) {
+        resolve();
+        cxpcStore.offPricesReceived(check);
+      }
+    };
+    cxpcStore.onPricesReceived(check);
   });
 }